+++ integration/sample/src/public_api.ts
@@ -1,2 +1,3 @@
-export * from './bar';
+export * from './bar/bar.component';
 export * from './foo/foo.component';
+export * from './ui-lib.module';