+++ integration/sample/src/ui-lib.module.ts
@@ -1,0 +1,16 @@
+import { NgModule } from '@angular/core';
+import { FooComponent } from './foo/foo.component';
+import { BarComponent } from './bar/bar.component';
+
+
+@NgModule({
+  declarations: [
+    FooComponent,
+    BarComponent
+  ],
+  exports: [
+    FooComponent,
+    BarComponent
+  ]
+})
+export class UiLibModule {}